Added Listener Interface
All classes which can be added to the EventBus's <listener> array must now implement the Listener interface.  This interface just requires that all classes which implement it must have an implementation of the catchEvent() method.

This is mostly just to make the code safer.  It's no longer possible for the event bus to try to call catchEvent() on a class which does not implement it.