Finish Weather
Added the last needed functionality to the Weather package:

Weather.loadWeathersForAllLocations() returns an arraylist containing Weather objects for every enumerated location.

getXWindSpeed() and getYWindSpeed() return the orthogonal components of the wind speed.

Also added a few comments and moved the getters/setters to the bottom to follow the style guide.